make windows tests work

+++ tests/test_installer.sh
@@ -2,21 +2,32 @@
 
 set -e
 
+if [ -z "TEST_REPO_TOKEN" ]; then
+  echo "TEST_REPO_TOKEN must be set"
+  exit -1
+fi
 SCRIPTDIR=$(cd $(dirname $BASH_SOURCE[0]) && pwd)
-vflag=$1; shift
+CONDA_PREFIX=$(cd $CONDA_PREFIX && pwd)
+if [ ! -z $1 ]; then vflag="==$1"; shift; fi
 source $CONDA_PREFIX/*/activate
+# Needed to convert windows path to unix
+CONDA_PREFIX=$(cd $CONDA_PREFIX && pwd)
 
 SITE=https://repo.anaconda.cloud
 ALIAS=$SITE/repo
-anaconda-keymgr --version 999 --repo-token $TEST_REPO_TOKEN \
-    --default-channel $ALIAS/main --default-channel $ALIAS/msys2 \
-    --config-string full:installertest
+python -m anaconda_ident.keymgr --version 999 \
+    --repo-token $TEST_REPO_TOKEN --config-string full:installertest \
+    --default-channel $ALIAS/main --default-channel $ALIAS/msys2
 mkdir -p $CONDA_PREFIX/conda-bld/noarch
 mv anaconda-ident-config-999-default_0.tar.bz2 $CONDA_PREFIX/conda-bld/noarch
 python -m conda_index $CONDA_PREFIX/conda-bld
 
-TMPDIR=$(mktemp -d)
-trap 'rm -rf -- "$TMPDIR"' EXIT
+if [ -z "$TMPDIR" ]; then
+  TMPDIR=$(mktemp -d)
+  trap 'rm -rf -- "$TMPDIR"' EXIT
+else
+  TMPDIR=$(cd $TMPDIR && pwd)
+fi
 echo $TMPDIR
 cd $TMPDIR
 
@@ -55,7 +66,28 @@ echo "-----"
 
 constructor .
 
-T_PREFIX=$TMPDIR/aidtest
-[ -f AIDTest*.sh ] || exit 0
-bash AIDTest*.sh -b -p $T_PREFIX -k
+if [ -f AIDTest*.sh ]; then
+  T_PREFIX=$TMPDIR/aidtest
+  echo "Running the .sh installer..."
+  bash AIDTest*.sh -b -p $T_PREFIX -k
+elif [ -f AIDTest*.exe ]; then
+  echo "Running the .exe installer..."
+  cmd.exe /c "$(ls -1 AIDTest*.exe) /S"
+  echo "Installer has been run"
+  W_PREFIX="$USERPROFILE\\aidtest"
+  T_PREFIX=$(cd "$W_PREFIX" && pwd)
+else
+  echo "No installer created"
+  exit -1
+fi
 . $SCRIPTDIR/test_environment.sh "$T_PREFIX" "$TEST_REPO_TOKEN"
+if [ ! -z "$W_PREFIX" ]; then
+  echo "Running the .exe uninstaller"
+  cmd.exe /c "$W_PREFIX\\Uninstall-AIDTest.exe /S"
+fi
+echo "success .. $success"
+if [ "$success" = yes ]; then
+  exit 0
+else
+  exit -1
+fi
